Rename VerticalDatumCorrections to SurfaceCorrections.
